add handlers `list`, `get`
