+++ internal/service/metricservice/metricservice.go
@@ -11,6 +11,8 @@ import (
 type (
 	MetricService interface {
 		Push(name, kind, value string) error
+		Get(name, kind string) (string, error)
+		GetAll() []storage.Record
 	}
 
 	metricService struct {
@@ -50,4 +52,25 @@ func (s metricService) Push(name, kind, value string) error {
 	fmt.Println("name:", name, record)
 
 	return s.storage.Push(name, record)
+}
+
+func (s metricService) Get(name, kind string) (string, error) {
+	_, err := metric.GetKind(kind)
+	if err != nil {
+		return "", err
+	}
+
+	record, ok := s.storage.Get(name)
+	if !ok {
+		return "", metric.ErrorMetricNotFound
+	}
+
+	value := record.GetValue().String()
+	return value, nil
+}
+
+func (s metricService) GetAll() []storage.Record {
+	records := s.storage.GetAll()
+
+	return records
 }
